+++ applications/sentiment_analysis/predict.py
@@ -55,8 +55,8 @@ def predict_ext(ext_model_path, ext_label_path, test_path, max_seq_len):
     test_ds = copy.copy(ori_test_ds).map(trans_func, lazy=False)
 
     batchify_fn = lambda samples, fn=Tuple(
-        Pad(axis=0, pad_val=tokenizer.pad_token_id),
-        Pad(axis=0, pad_val=tokenizer.pad_token_type_id),
+        Pad(axis=0, pad_val=tokenizer.pad_token_id, dtype="int64"),
+        Pad(axis=0, pad_val=tokenizer.pad_token_type_id, dtype="int64"),
         Stack(dtype="int64"), ): fn(samples)
 
     test_batch_sampler = paddle.io.BatchSampler(
